backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi
uvicorn[standard]
python-multipart
opencv-python-headless
numpy
pillow
ultralytics
python-dotenv
torch==2.3.1+cpu
torchvision==0.18.1+cpu
